+++ packages/blob-stores/CHANGELOG.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-
+  - BLOB stores: new `storeSize` `Hourly` - as `Daily` but with sub-folder for each hour inside a day folder
+  - BLOB sores: new property `LUCount` - a count of Logical Units BLOB stores divided into.
+    If > 0 then files are stored inside `Logical Unit` sub-folders `/LU01`, `/LU02`,.
+     
+    Write operations works with last LU folder. Each LU folder can be mounted to his own partition.
+    In this case `tempPath` should point to the same partition as last LU.
+    
 ### Changed
 
 ### Deprecated